all done

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: epakdama <epakdama@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/16 11:29:38 by epakdama          #+#    #+#             */
-/*   Updated: 2025/06/16 15:34:35 by epakdama         ###   ########.fr       */
+/*   Updated: 2025/06/16 15:47:04 by epakdama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,12 +25,15 @@ static void	write_variadic(char *s, int i, va_list args, int *len)
 	else if (s[i + 1] == 'p')
 	{
 		ft_putstr("0x", len);
-		ft_putnbr_base(va_arg(args, unsigned int), "0123456789abcdef", len);
+		ft_putnbr_base(va_arg(args, unsigned long long), "0123456789abcdef",
+			len);
 	}
 	else if (s[i + 1] == 'x')
-		ft_putnbr_base(va_arg(args, unsigned int), "0123456789abcdef", len);
+		ft_putnbr_base(va_arg(args, unsigned long long), "0123456789abcdef",
+			len);
 	else if (s[i + 1] == 'X')
-		ft_putnbr_base(va_arg(args, unsigned int), "0123456789ABCDEF", len);
+		ft_putnbr_base(va_arg(args, unsigned long long), "0123456789ABCDEF",
+			len);
 }
 
 int	ft_printf(const char *s, ...)
